Improved implementation of CreateMemo Fixed implementation of CompileAppendToList for interfaces Fixed missing ROCOCO_API on allocators in sexy.global.functions.cpp Improved messaging for the mhost gui visual logger T3D now gives an improved display for the rolling logger

+++ source/rococo/rococo.mplat/mplat.graphics.materials.cpp
@@ -158,7 +158,7 @@ namespace Rococo::Graphics
 
 		  for (size_t i = 0; i < metrics.NumberOfElements; ++i)
 		  {
-			  auto id = (MaterialId)i;
+			  auto id = MaterialId((int32)i);
 			  cstr name = renderer.Materials().GetMaterialTextureName(id);
 			  if (name)
 			  {
